ls_15 products service added
