+++ store/src/app/products/products.component.ts
@@ -1,34 +1,26 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Injectable, OnInit } from '@angular/core';
 import { Product } from '../interfaces/products.interface';
+import { ProductsService } from './products.service';
 
 @Component({
   selector: 'app-products',
   templateUrl: './products.component.html',
-  styleUrls: ['./products.component.scss']
+  styleUrls: ['./products.component.scss'],
+  providers: [ProductsService]
 })
 
+@Injectable()
 export class ProductsComponent implements OnInit {
+
+  constructor(private productsService: ProductsService) { };
 
   products: Product[] = [];
 
   ngOnInit() {
-    this.products = this.generateProducts(8)
+    this.getProducts(8);
   }
 
-  generateProducts(n: number): Product[] {
-    const names: string[] = ['Xiaomi 12', 'AirPods', 'Iphone 14', 'Asus ROG 17',
-      'Mi AirDots', 'Sony WH-1000XM4', 'Power Bank 20000', 'Invertor 12-220',
-      'BFG 9000', 'Doom Eternal'];
-    let arr: Product[] = [];
-
-    for (let i = 0; i < n; i++) {
-      const obj = {
-        id: i,
-        name: names[Math.floor(Math.random() * 10)],
-        price: Math.floor(Math.random() * 100)
-      };
-      arr.push(obj);
-    }
-    return arr;
+  getProducts(n: number): void {
+    this.products = this.productsService.generateProducts(n);
   }
 }
